Added namespace EventHandlers and split methods from Bot to specific classes
